fix: SNT-392 fix middlewares (#2857)
refs: SNT-392

* fix middlewares

* replace QueryCountMiddleware with SafeQueryCountMiddleware
