+++ iaso/middlewares/query_count.py
@@ -1,22 +1,6 @@
 from collections import defaultdict
 
-from django.urls import Resolver404, resolve
-from djangorestframework_camel_case.middleware import CamelCaseMiddleWare
 from querycount.middleware import QueryCountMiddleware
-
-
-class CustomCamelCaseMiddleWare(CamelCaseMiddleWare):
-    def __call__(self, request):
-        # we only execute this middleware on profiles API at the moment
-        route = ""
-        try:
-            route = resolve(request.path_info).route
-        except Resolver404:
-            pass
-
-        if route.startswith("/api/v2/profiles"):
-            return super().__call__(request)
-        return self.get_response(request)
 
 
 class SafeQueryCountMiddleware(QueryCountMiddleware):
